ISU_RESUME/CS402/ug_sb_4/Backend/requirements.txt
docxtpl==0.20.2
Flask==3.1.2
numpy==2.3.3
openpyxl==3.1.5
pandas==3.0.1
pypdf==6.10.0
python-docx==1.2.0
RapidFuzz==3.14.5